requirements.txt
Flask
Pillow
easyocr
flask-swagger-ui
